# back-end/luego-config-service/artefacts/domains/Configuration/cluster-config-demo/1.0.0/openapi-spec.yaml
openapi: 3.1.0
info:
  title: cluster-config-demo
  version: 1.0.0
paths:
...
  /v1/domains/Configuration/apps/cluster-config-demo/1.0.0/models/demo.config.configureKafkaCluster/configure:
    post:
      tags:
      - demo.config - Configuration
      summary: Configure a Kafka cluster by determining the best type of cluster and
        configuring its details.
      description: Configure a Kafka cluster by determining the best type of cluster
        and configuring its details.
      operationId: decision:demo.config.configureKafkaCluster
      parameters:
      - name: lang
        in: query
        required: false
        schema:
          type: string
          enum:
          - en
          - fr
      - name: richResults
        in: query
        required: false
        schema:
          type: boolean
          enum:
          - true
          - false
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                the customer request:
                  $ref: '#/components/schemas/demo.config.CustomerRequest'
                the configuration:
                  $ref: '#/components/schemas/demo.config.Configuration'
              additionalProperties: false
        required: true
      responses:
        '200':
          description: Ok, response generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  computationDetails:
                    type: object
                    properties:
                      totalTime:
                        type: number
                      appName:
                        type: string
                      appVersion:
                        type: string
                      operation:
                        type: string
                      cycles:
                        type: number
                  missingData:
                    type: object
                  result:
                    type: object
                    properties:
                      the customer request:
                        $ref: '#/components/schemas/demo.config.CustomerRequest'
                      the configuration:
                        $ref: '#/components/schemas/demo.config.Configuration'
                    additionalProperties: false
                additionalProperties: false
        '400':
          description: Something went wrong
...
  /v1/domains/Configuration/apps/cluster-config-demo/1.0.0/models/demo.config.configureKafkaCluster/infer:
    post:
      tags:
      - demo.config - Inference (for test purposes)
      summary: Configure a Kafka cluster by determining the best type of cluster and
        configuring its details.
      description: Configure a Kafka cluster by determining the best type of cluster
        and configuring its details.
      operationId: decision:demo.config.configureKafkaCluster
      parameters:
      - name: lang
        in: query
        required: false
        schema:
          type: string
          enum:
          - en
          - fr
      - name: richResults
        in: query
        required: false
        schema:
          type: boolean
          enum:
          - true
          - false
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                the customer request:
                  $ref: '#/components/schemas/demo.config.CustomerRequest'
                the configuration:
                  $ref: '#/components/schemas/demo.config.Configuration'
              additionalProperties: false
        required: true
      responses:
        '200':
          description: Ok, response generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/luego.results.UnitResponse'
        '400':
          description: Something went wrong
...
  /v1/domains/Configuration/apps/cluster-config-demo/1.0.0/models/demo.config.configureKafkaCluster/initial_payload:
    get:
      tags:
      - demo.config - Payloads
      operationId: decision:demo.config.configureKafkaCluster
      parameters:
      - name: lang
        in: query
        required: false
        schema:
          type: string
          enum:
          - en
          - fr
      - name: richResults
        in: query
        required: false
        schema:
          type: boolean
          enum:
          - true
          - false
      responses:
        '200':
          description: Ok, response generated
          content:
            application/json: {}
        '400':
          description: Something went wrong
...
components:
  schemas:
    demo.config.ScalingType:
      type: string
      enum:
      - Elastic
      - Manual
    demo.config.SubComponentCode:
      type: string
      pattern: '[ABC][123]'
    demo.config.CloudProvider:
      type: string
      enum:
      - AWS
      - Azure
      - GCP
      - CP4
      - CP5
    demo.config.ZeroTo1200:
      type: number
      minimum: 0.0
      maximum: 1200.0
    demo.config.ZeroTo2400:
      type: number
      minimum: 0.0
      maximum: 2400.0
    demo.config.Email:
      type: string
      maxLength: 36
      minLength: 8
      pattern: ^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+$
    demo.config.CustomerRequest:
      type: object
      required:
      - LGType_
      properties:
        LGType_:
          type: string
          enum:
          - demo.config.CustomerRequest
        cloudProvider:
          $ref: '#/components/schemas/demo.config.CloudProvider'
        mustHaveDedicatedEnv:
          type: boolean
        mustHavePrivateEndpoints:
          type: boolean
        mustHaveSubSecondLatency:
          type: boolean
        isProductionUseCase:
          type: boolean
        ingress:
          type: number
          format: double
        egress:
          type: number
          format: double
        throughput:
          type: number
          format: double
        numberOfConnections:
          type: integer
          format: int32
      additionalProperties: false
    demo.config.Configuration:
      type: object
      required:
      - LGType_
      properties:
        LGType_:
          type: string
          enum:
          - demo.config.Configuration
        cluster:
          $ref: '#/components/schemas/demo.config.Cluster'
      additionalProperties: false
    demo.config.Cluster:
      type: object
      oneOf:
      - $ref: '#/components/schemas/demo.config.StandardCluster'
      - $ref: '#/components/schemas/demo.config.DedicatedCluster'
      - $ref: '#/components/schemas/demo.config.EnterpriseCluster'
      - $ref: '#/components/schemas/demo.config.BasicCluster'
      - $ref: '#/components/schemas/demo.config.FreightCluster'
      required:
      - LGType_
      discriminator:
        propertyName: LGType_
      properties:
        LGType_:
          type: string
          enum:
          - demo.config.StandardCluster
          - demo.config.DedicatedCluster
          - demo.config.EnterpriseCluster
          - demo.config.BasicCluster
          - demo.config.FreightCluster
        scaling:
          $ref: '#/components/schemas/demo.config.ScalingType'
        minNumberOfNodes:
          type: integer
          format: int32
      additionalProperties: false
    demo.config.StandardCluster:
      type: object
      required:
      - LGType_
      discriminator:
        propertyName: LGType_
      properties:
        LGType_:
          type: string
          enum:
          - demo.config.StandardCluster
        subComponents:
          type: array
          minItems: 1
          items:
            type: object
            required:
            - LGType_
            properties:
              LGType_:
                type: string
                enum:
                - demo.config.SubComponent
              code:
                $ref: '#/components/schemas/demo.config.SubComponentCode'
              capacity:
                $ref: '#/components/schemas/demo.config.ZeroTo2400'
              subSubComponents:
                type: array
                items:
                  type: object
                  required:
                  - LGType_
                  properties:
                    LGType_:
                      type: string
                      enum:
                      - demo.config.SubSubComponent
                    description:
                      type: string
                  additionalProperties: false
            additionalProperties: false
      additionalProperties: false
    demo.config.DedicatedCluster:
      type: object
      required:
      - LGType_
      discriminator:
        propertyName: LGType_
      properties:
        LGType_:
          type: string
          enum:
          - demo.config.DedicatedCluster
        optionalComponent:
          type: array
          maxItems: 1
          minItems: 0
          items:
            type: object
            oneOf:
            - $ref: '#/components/schemas/demo.config.OptionA'
            - $ref: '#/components/schemas/demo.config.OptionB'
            required:
            - LGType_
            discriminator:
              propertyName: LGType_
            properties:
              LGType_:
                type: string
                enum:
                - demo.config.OptionA
                - demo.config.EmailNotification
                - demo.config.OptionB2
            additionalProperties: false
      additionalProperties: false
    demo.config.EnterpriseCluster:
      type: object
      required:
      - LGType_
      discriminator:
        propertyName: LGType_
      properties:
        LGType_:
          type: string
          enum:
          - demo.config.EnterpriseCluster
        optionalComponent:
          type: array
          maxItems: 1
          minItems: 0
          items:
            type: object
            oneOf:
            - $ref: '#/components/schemas/demo.config.OptionA'
            - $ref: '#/components/schemas/demo.config.OptionB'
            required:
            - LGType_
            discriminator:
              propertyName: LGType_
            properties:
              LGType_:
                type: string
                enum:
                - demo.config.OptionA
                - demo.config.EmailNotification
                - demo.config.OptionB2
            additionalProperties: false
      additionalProperties: false
    demo.config.BasicCluster:
      type: object
      required:
      - LGType_
      discriminator:
        propertyName: LGType_
      properties:
        LGType_:
          type: string
          enum:
          - demo.config.BasicCluster
      additionalProperties: false
    demo.config.FreightCluster:
      type: object
      required:
      - LGType_
      discriminator:
        propertyName: LGType_
      properties:
        LGType_:
          type: string
          enum:
          - demo.config.FreightCluster
      additionalProperties: false
...
    demo.config.OptionA:
      type: object
      required:
      - LGType_
      discriminator:
        propertyName: LGType_
      properties:
        LGType_:
          type: string
          enum:
          - demo.config.OptionA
        quantity:
          $ref: '#/components/schemas/demo.config.ZeroTo1200'
      additionalProperties: false
    demo.config.OptionB:
      type: object
      oneOf:
      - $ref: '#/components/schemas/demo.config.EmailNotification'
      - $ref: '#/components/schemas/demo.config.OptionB2'
      required:
      - LGType_
      discriminator:
        propertyName: LGType_
      properties:
        LGType_:
          type: string
          enum:
          - demo.config.EmailNotification
          - demo.config.OptionB2
      additionalProperties: false
    demo.config.EmailNotification:
      type: object
      required:
      - LGType_
      discriminator:
        propertyName: LGType_
      properties:
        LGType_:
          type: string
          enum:
          - demo.config.EmailNotification
        email:
          $ref: '#/components/schemas/demo.config.Email'
      additionalProperties: false
    demo.config.OptionB2:
      type: object
      required:
      - LGType_
      discriminator:
        propertyName: LGType_
      properties:
        LGType_:
          type: string
          enum:
          - demo.config.OptionB2
        quantity:
          $ref: '#/components/schemas/demo.config.ZeroTo2400'
      additionalProperties: false
...
    luego.results.UnitResponse:
      type: object
      properties:
        computationDetails:
          type: object
          properties:
            totalTime:
              type: number
            appVersion:
              type: string
        result:
          type: object
...